Add rest of controllers, implement endpoints for basic user views on frontend

+++ Controllers/BrandController.cs
@@ -1,8 +1,10 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using PizzaApp.Data;
 using PizzaApp.DTOs;
 using PizzaApp.Entities;
+using PizzaApp.Interfaces;
 
 namespace PizzaApp.Controllers
 {
@@ -11,42 +13,16 @@ namespace PizzaApp.Controllers
     public class BrandController : ControllerBase
     {
         private readonly AppDbContext _context;
+        private readonly IUserContextService _userContextService;
 
-        public BrandController(AppDbContext context)
+        public BrandController(AppDbContext context, IUserContextService userContextService)
         {
             _context = context;
+            _userContextService = userContextService;
         }
 
-        // POST: api/Brand
-        [HttpPost]
-        public async Task<ActionResult> CreateBrand([FromBody] CreateBrandDto dto)
-        {
-            var ownerExists = await _context.Owners.AnyAsync(o => o.Id == dto.OwnerId);
-            if (!ownerExists)
-            {
-                return BadRequest("Podany właściciel nie istnieje.");
-            }
-
-            if (await _context.Brands.AnyAsync(b => b.Name == dto.Name))
-            {
-                return Conflict($"Marka o nazwie '{dto.Name}' już istnieje.");
-            }
-
-            var brand = new Brand
-            {
-                Name = dto.Name,
-                Logo = dto.Logo,
-                OwnerId = dto.OwnerId
-            };
-
-            _context.Brands.Add(brand);
-            await _context.SaveChangesAsync();
-
-            return CreatedAtAction(nameof(GetBrand), new { id = brand.Id }, new BrandDto { Id = brand.Id, Name = brand.Name });
-        }
-
-        // GET: api/Brand
-        [HttpGet]
+        // GET: api/Brand/GetAll
+        [HttpGet("GetAll")]
         public async Task<ActionResult<IEnumerable<BrandDto>>> GetBrands()
         {
             var brands = await _context.Brands
@@ -59,6 +35,7 @@ namespace PizzaApp.Controllers
 
             return Ok(brands);
         }
+
 
         // GET: api/Brand/{id}
         [HttpGet("{id}")]
@@ -80,25 +57,72 @@ namespace PizzaApp.Controllers
                 Id = brand.Id,
                 Name = brand.Name,
                 Logo = brand.Logo,
-                Pizzerias = brand.Pizzerias.Select(p => new PizzeriaSimpleDto
+                Pizzerias = [.. brand.Pizzerias.Select(p => new PizzeriaSimpleDto
                 {
                     Id = p.Id,
                     Name = p.Name,
                     City = p.Address?.City?.Name ?? "Nieznane"
-                }).ToList()
+                })]
             };
 
             return Ok(details);
         }
 
+
+        // POST: api/Brand/
+        [HttpPost]
+        [Authorize]
+        public async Task<ActionResult> CreateBrand([FromBody] CreateBrandDto dto)
+        {
+            var userId = _userContextService.GetUserId();
+
+            if(!userId.HasValue)
+            {
+                return Unauthorized();
+            }
+            var ownerExists = await _context.Owners.AnyAsync(o => o.Id == userId);
+
+            if (!ownerExists)
+            {
+                return BadRequest("Podany właściciel nie istnieje.");
+            }
+
+            if (await _context.Brands.AnyAsync(b => b.Name == dto.Name))
+            {
+                return Conflict($"Marka o nazwie '{dto.Name}' już istnieje.");
+            }
+
+            var brand = new Brand
+            {
+                Name = dto.Name,
+                Logo = dto.Logo,
+                OwnerId = userId.Value
+            };
+
+            _context.Brands.Add(brand);
+            await _context.SaveChangesAsync();
+
+            return CreatedAtAction(nameof(GetBrand), new { id = brand.Id }, new BrandDto { Id = brand.Id, Name = brand.Name });
+        }
+
         // DELETE: api/Brand/{id}
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> DeleteBrand(Guid id)
         {
+            var userId = _userContextService.GetUserId();
+            if (!userId.HasValue)
+            {
+                return Unauthorized();
+            }
             var brand = await _context.Brands.FindAsync(id);
             if (brand == null)
             {
                 return NotFound("Marka nie istnieje.");
+            }
+            if(brand.Owner!.Id != userId.Value)
+            {
+                return Forbid();
             }
 
             _context.Brands.Remove(brand);
